# getindex for CuArray to fix repeated indices error