--- a/src/lib/broadcast.jl
+++ b/src/lib/broadcast.jl
@@ -288,16 +288,9 @@ using GPUArraysCore  # replaces @require CUDA block, weird indenting to preserve
 
   ∇getindex(x::CUDA.CuArray, inds::Tuple{AbstractArray{<:Integer}}) = dy -> begin
     inds1_cpu = Array(inds[1])
-    if allunique(inds1_cpu)
-      dx = _zero(x, eltype(dy))
-      dxv = view(dx, inds[1])
-      dxv .= accum.(dxv, _droplike(dy, dxv))
-      return _project(x, dx), nothing
-    else
-      dx = zeros(eltype(dy), length(x))
-      dxv = view(dx, inds1_cpu)
-      dxv .= accum.(dxv, _droplike(Array(dy), dxv))
-      return _project(x, CUDA.CuArray(dx)), nothing
-    end
+    dx = zeros(eltype(dy), length(x))
+    dxv = view(dx, inds1_cpu)
+    dxv .= accum.(dxv, _droplike(Array(dy), dxv))
+    return _project(x, CUDA.CuArray(dx)), nothing
   end
 end