I spent quite a while working out where my gradients were wrong, until eventually tracking it down to the GPU repeated index issue that is mentioned in multiple places:
- https://github.com/FluxML/Zygote.jl/issues/600
- https://github.com/FluxML/Zygote.jl/issues/821
- https://github.com/JuliaGPU/CUDA.jl/issues/89
- https://github.com/JuliaGPU/CUDA.jl/issues/1019
- https://github.com/JuliaLang/julia/pull/31407

Based on a solution suggested at https://discourse.julialang.org/t/increment-elements-of-array-by-index/49694 I added a hack to get the correct gradients, which can be found in this PR. It uses `scatter!` from [NNlib](https://github.com/FluxML/NNlib.jl) to accumulate values over multiple indices. It is considerably slower than the current version, which on the GPU gives silently incorrect and even stochastic gradients. However it is faster than moving the array off the GPU to do this step.

This isn't currently suitable for merging since it adds dependencies and isn't tested beyond the simple case. It might be useful as a starting point for a discussion on how to get this important issue fixed, though, and also might be useful as a hack to others.