@CarloLucibello 2021-12-08
Since Zygote doesn't depend on CUDA and NNlib I'm not sure where we should put this. Here but inside a `@require`? In `NNlib/NNlibCUDA`?