@jgreener64 2021-12-08
> Maybe taking a fast path after a "has duplicates" check can help performance a bit?

Yes I would imagine so.

> I'm not sure where we should put this

There is already a CUDA require block in `lib/broadcast.jl` though that still leaves the NNlib dependency. Another option is to rewrite a simple version of scatter here.