@ToucheSir 2021-12-08
Is this something ReverseDiff could benefit from as well? If so, we could consider adding it as an `rrule` with `RuleConfig{>:HasReverseMode}` to NNlibCUDA.