@DhairyaLGandhi 2021-12-09
I don't think this is something that can be solved well with an adjoint at this point. Specifically, note https://github.com/JuliaGPU/CUDA.jl/issues/89 which describes why this happens. Best fixed at the CUDA.jl level based on how it is implemented.