@CarloLucibello 2021-12-25
Can we have some benchmarks here?