@ToucheSir 2021-12-25
So the main concern wrt. dependencies here is not CUDA (because there's already a `Requires` block for it in Zygote) but NNlib(CUDA). I think one way to make this more palatable would be to extract `scatter!(+, dx, dy, inds)` into a kernel and create a wrapper accumulation function that dispatches to this kernel when passed CuArrays. The default fallback could reuse the code in https://github.com/FluxML/Zygote.jl/blob/v0.6.33/src/lib/array.jl#L38-L39.