@CarloLucibello 2022-02-12
Yes, we could do that, not ideal but seems the best alternative