@mcabbott 2022-02-20
Do these times need a `CUDA.@sync`? Seems a little crazy that copying to an Array could be quicker, but if it really is then that's simple. And if so, it should probably skip the check? `allunique` is surprisingly slow, has to make a Dict I think:
```
julia> @btime allunique($(shuffle!(collect(1:1000))));
  min 11.958 μs, mean 20.876 μs (17 allocations, 49.14 KiB)
```