@mcabbott 2022-08-03
https://github.com/JuliaDiff/ChainRules.jl/pull/655 implements the same copy-to-CPU idea is [98e87c3](https://github.com/JuliaDiff/ChainRules.jl/pull/655/commits/98e87c3e7a560ce094969be6df49e569d37fc61a). Comments (or benchmarking) would be welcome.

Adding test cases here would also be a great idea, as they will be run on a real GPU via buildkite.