@jgreener64 2023-10-06
I think this was fixed in https://github.com/JuliaDiff/ChainRules.jl/pull/655.